+++ lib/src/models/candle.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 /// Candle model which holds a single candle data.
 /// It contains five required double variables that hold a single candle data: high, low, open, close and volume.
 /// It can be instantiated using its default constructor or fromJson named constructor.
@@ -22,6 +24,12 @@ class Candle {
   /// Volume is the number of shares of a
   /// security traded during a given period of time.
   final double volume;
+
+  double? ma7, ma25, ma99;
+
+  double? get maxMa => ma7 != null && ma25 != null && ma99 != null ? max(max(ma7 ?? 0, ma25 ?? 0), ma99 ?? 0) : null;
+
+  double? get minMa => ma7 != null && ma25 != null && ma99 != null ? min(min(ma7 ?? 0, ma25 ?? 0), ma99 ?? 0) : null;
 
   bool get isBull => open <= close;